Move createComplexityRule out of index file #65
