+++ src/index.ts
@@ -1,21 +1,10 @@
 /**
  * Created by Ivo Meißner on 28.07.17.
- *
- * @flow
  */
-import { ValidationContext } from 'graphql';
-import QueryComplexity from './QueryComplexity.js';
-import { QueryComplexityOptions } from './QueryComplexity.js';
 
 export * from './estimators/index.js';
 export * from './QueryComplexity.js';
+import { createComplexityRule as createComplexityRuleFn } from './createComplexityRule.js';
 
-export function createComplexityRule(
-  options: QueryComplexityOptions
-): (context: ValidationContext) => QueryComplexity {
-  return (context: ValidationContext): QueryComplexity => {
-    return new QueryComplexity(context, options);
-  };
-}
-
+export const createComplexityRule = createComplexityRuleFn;
 export default createComplexityRule;
